split db.py
