+++ tests/unit/test_load_db.py
@@ -1,6 +1,7 @@
 from freezegun import freeze_time
 from gudlft.models.dataloader import DataLoader
-from gudlft.models.db import Club, Competition
+from gudlft.models.competition import Competition
+from gudlft.models.club import Club
 
 
 class TestLoadDb:
